improved newsletter functionality
